+++ tests/test-bootstrap.php
@@ -64,4 +64,23 @@ class BootstrapTest extends WP_UnitTestCase {
 		$this->assertEqualSetsWithIndex($expected, $actual);
 	}
 
+
+
+	/**
+	 * Gist authenticates as GitHub, so its hosts join the github set.
+	 */
+	public function test_set_credential_hosts_adds_gist_hosts_to_github_set() {
+		$hosts = (new Bootstrap())->set_credential_hosts([]);
+
+		$this->assertContains('gist.github.com', $hosts['github']);
+		$this->assertContains('gist.githubusercontent.com', $hosts['github']);
+	}
+
+	public function test_load_hooks_registers_credential_hosts_filter() {
+		$bootstrap = new Bootstrap();
+		$bootstrap->load_hooks();
+
+		$this->assertNotFalse(has_filter('gu_credential_hosts', [$bootstrap, 'set_credential_hosts']));
+	}
+
 }